data-pipeline/pyproject.toml
[project]
name = "rld-data-pipeline"
version = "0.1.0"
description = "Aave V3 data pipeline extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.2",
    "web3>=6.15.0",
    "asyncpg>=0.29.0",
]
